修改了双色球大额投注失败的问题

+++ Ruyicai_General/v3.10.1/src/com/ruyicai/activity/buy/miss/AddViewMiss.java
@@ -313,8 +313,8 @@ public class AddViewMiss {
 		return allZhu;
 	}
 
-	public int getAllAmt() {
-		int allAmt = 0;
+	public long getAllAmt() {
+		long allAmt = 0;
 		for (CodeInfo codeInfo : codeList) {
 			allAmt += codeInfo.getAmt();
 		}
